Put misc functions in separate file
Remove some more clutter from goaci.go.
